requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
Django==4.0
django-appconf==1.0.5
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-imagekit==4.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-nested-inline==0.4.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pilkit==2.0
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
